+++ internal/cli/serverless/export/download.go
@@ -36,7 +36,7 @@ import (
 	"tidbcloud-cli/internal/service/cloud"
 	"tidbcloud-cli/internal/ui"
 	"tidbcloud-cli/internal/util"
-	exportClient "tidbcloud-cli/pkg/tidbcloud/v1beta1/serverless_export"
+	exportClient "tidbcloud-cli/pkg/tidbcloud/v1beta1/serverless/export"
 )
 
 const DefaultConcurrency = 3
