+++ pkg/tidbcloud/v1beta1/serverless/export/docs/ExportCompressionTypeEnum.md
