+++ pkg/tidbcloud/v1beta1/serverless/export/docs/ExportS3AuthTypeEnum.md
