+++ pkg/tidbcloud/v1beta1/serverless/export/model_export_compression_type_enum.go
