+++ pkg/tidbcloud/v1beta1/serverless/export/model_export_s3_auth_type_enum.go
